+++ qord/dataclasses/files.py
@@ -42,19 +42,16 @@ class File:
     Example usage with :meth:`~BaseMessageChannel.send` method::
 
         file = qord.File("path/to/file.png")
-        await channel.send(files=[file])
+        await channel.send(file=file)
 
     Parameters
     ----------
     content: Union[:class:`builtins.str`, :class:`builtins.bytes`, :class:`io.BufferedReader`]
         The content of files. If a string is being passed, It would be considered
         the file path and would be opened and red in "read binary" mode.
-
-        When passing a :class:`builtins.bytes` object, The ``name`` parameter
-        must be included.
     name: :class:`builtins.str`
-        The name of file. This parameter is required when passing :class:`builtins.bytes`
-        in ``content`` parameter.
+        The name of file. The name would be retrieved from given file path or buffer
+        object if possible and would fallback to "untitled" if couldn't be resolved.
     spoiler: :class:`builtins.bool`
         Whether the file should be marked as spoiler when sent.
     description: :class:`builtins.str`
@@ -77,8 +74,8 @@ class File:
         content: typing.Union[str, bytes, BufferedReader],
         /,
         *,
-        name: str = UNDEFINED,
-        description: str = UNDEFINED,
+        name: typing.Optional[str] = None,
+        description: typing.Optional[str] = None,
         spoiler: bool = False,
     ) -> None:
 
@@ -86,13 +83,10 @@ class File:
             with open(content, "rb") as f:
                 self.content = f.read()
 
-            if name is UNDEFINED:
+            if name is None:
                 name = os.path.basename(content)
 
         elif isinstance(content, bytes):
-            if name is UNDEFINED:
-                raise TypeError("name parameter must be passed when passing in bytes object.")
-
             self.content = content
 
         elif isinstance(content, BufferedReader):
@@ -101,11 +95,11 @@ class File:
 
             self.content = content.read()
 
-            if name is UNDEFINED:
+            if name is None:
                 name = content.name
 
-        if name is UNDEFINED:
-            raise RuntimeError("Could not resolve the file name, probably invalid type.")
+        if name is None:
+            name = "untitled"
 
         self.name = name
         self.description = description